修改整数溢出问题
# message auto-generated for no-merge-commit merge:
merge master_fix into master

修改整数溢出问题

Created-by: insightDeveloper

Match-id-ae44d2266280f6e4f7c6d9bc4873516515f2820a

+++ server/msinsight/src/utils/FileUtil.h
@@ -353,7 +353,7 @@ public:
         std::string rankId = GetRankIdFromPath(timeLineFile);
         // 上上层目录名没有则取文件的名字
         if (rankId.empty()) {
-            int index3 = timeLineFile.find_last_of('.');
+            size_t index3 = timeLineFile.find_last_of('.');
             rankId = timeLineFile.substr(parent.length() + 1, index3 - parent.length() - 1);
         }
         return rankId;
